functions/store_data/requirements.txt
functions-framework==3.5.0
google-cloud-firestore
pusher